refactored "rest" package to "api"
