+++ src/main/java/api/JAXRSConfig.java
@@ -1,4 +1,4 @@
-package rest;
+package api;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
