+++ src/main/java/api/resource/SubredditResource.java
@@ -1,4 +1,4 @@
-package rest.resource;
+package api.resource;
 
 import business.interfaces.ProfileService;
 import business.interfaces.SubredditService;
